Add seperate file for input details
