+++ main.rb
@@ -1,4 +1,5 @@
 require_relative 'app'
+require_relative 'input_details'
 
 @my_app = App.new
 
@@ -8,69 +9,6 @@ def main
   option = gets.chomp.to_i
 
   select_option(option)
-end
-
-def select_option(option)
-  case option
-  when 1..2
-    list_part(option)
-    main
-  when 3..5
-    create_part(option)
-    main
-  when 6
-    list_rental_by_person_id
-    main
-  when 7
-    puts 'Goodbye'
-  else
-    puts 'Choose a correct option'
-  end
-end
-
-def list_part(option)
-  case option
-  when 1
-    select_all_books
-    main
-  when 2
-    select_all_people
-    main
-  end
-end
-
-def create_part(option)
-  case option
-  when 3
-    select_person_input
-  when 4
-    input_book_details
-  when 5
-    input_rental_details
-  end
-end
-
-def select_all_books
-  @my_app.all_books.map { |book| puts "Title: #{book.title}, Author: #{book.author}" }
-end
-
-def select_all_people
-  @my_app.all_people.map { |person| puts "Name: #{person.name}, ID: #{person.id}, Age: #{person.age}" }
-end
-
-def select_person_input
-  print 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-  num = $stdin.gets.to_i
-  case num
-  when 1
-    input_student_details(num)
-  when 2
-    input_teacher_details(num)
-  else
-    print 'Input correct number: '
-    num = $stdin.gets.to_i
-    select_input(num)
-  end
 end
 
 def app_description
@@ -86,58 +24,14 @@ def app_description
   puts '7 - Exit'
 end
 
-def input_student_details(num)
-  puts 'Age: '
-  age = $stdin.gets.to_i
-  puts 'Name: '
-  name = $stdin.gets.chomp
-  puts 'Has parent permission? [Y/N]: '
-  permission = $stdin.gets
-  @my_app.add_person(num, age, name, permission)
-  puts 'Person successfully created'
-end
-
-def input_teacher_details(num)
-  puts 'Age: '
-  age = $stdin.gets.to_i
-  puts 'Name: '
-  name = $stdin.gets.chomp
-  puts 'Specialization: '
-  specialization = $stdin.gets.chomp
-  @my_app.add_person(num, age, name, specialization)
-  puts 'Person successfully created'
-end
-
-def input_book_details
-  puts 'Title: '
-  title = $stdin.gets.chomp
-  puts 'Author: '
-  author = $stdin.gets.chomp
-  @my_app.add_book(title, author)
-  puts 'Book successfully created'
-end
-
-def input_rental_details
-  puts 'Select a book from the following list by number (not id)'
-  @my_app.all_books.map.with_index { |book, index| puts "#{index}) Title: '#{book.title}', Author: #{book.author}" }
-  book_number = gets.chomp.to_i
-  puts ''
-  puts 'Select a person from the follwing list by number (not id)'
-  @my_app.all_people.map.with_index do |person, index|
-    puts "#{index}) Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+def select_option(option)
+  case option
+  when 7
+    puts 'Goodbye'
+    exit
+  else
+    puts 'Choose a correct option'
   end
-  person_number = gets.chomp.to_i
-  puts ''
-  puts 'Date: '
-  date = $stdin.gets
-  @my_app.add_rental(date, book_number, person_number)
-  puts 'Rental created succesfully'
 end
 
-def list_rental_by_person_id
-  puts 'ID of person: '
-  person_id = $stdin.gets.to_i
-  puts 'Rentals: '
-  @my_app.rental_by_id(person_id)
-end
 main
